Refactor CallbackChecker into test utils
