+++ tests/helpers/fixtures.py
@@ -37,13 +37,14 @@ import pytest
 import py.path  # pylint: disable=no-name-in-module
 
 import helpers.stubs as stubsmod
+from helpers.utils import CallbackChecker
 from qutebrowser.config import config, configdata, configtypes, configexc
 from qutebrowser.utils import objreg, standarddir
 from qutebrowser.browser.webkit import cookies
 from qutebrowser.misc import savemanager, sql
 from qutebrowser.keyinput import modeman
 
-from PyQt5.QtCore import pyqtSignal, QEvent, QSize, Qt, QObject
+from PyQt5.QtCore import QEvent, QSize, Qt
 from PyQt5.QtGui import QKeyEvent
 from PyQt5.QtWidgets import QWidget, QHBoxLayout, QVBoxLayout
 from PyQt5.QtNetwork import QNetworkCookieJar
@@ -76,34 +77,6 @@ class WinRegistryHelper:
     def cleanup(self):
         for win_id in self._ids:
             del objreg.window_registry[win_id]
-
-
-class CallbackChecker(QObject):
-
-    """Check if a value provided by a callback is the expected one."""
-
-    got_result = pyqtSignal(object)
-    UNSET = object()
-
-    def __init__(self, qtbot, parent=None):
-        super().__init__(parent)
-        self._qtbot = qtbot
-        self._result = self.UNSET
-
-    def callback(self, result):
-        """Callback which can be passed to runJavaScript."""
-        self._result = result
-        self.got_result.emit(result)
-
-    def check(self, expected):
-        """Wait until the JS result arrived and compare it."""
-        if self._result is self.UNSET:
-            with self._qtbot.waitSignal(self.got_result, timeout=2000):
-                pass
-        self._assert_result(self._result, expected)
-
-    def _assert_result(self, result, expected):
-        assert result == expected
 
 
 @pytest.fixture
